pyproject.toml
[project]
name = "lyume-memory-proxy"
version = "0.2.0"
description = "Drop-in memory proxy for local LLMs. Private. Learns from feedback."
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.0",
    "uvicorn>=0.42.0",
    "httpx>=0.28.0",
    "asyncpg>=0.31.0",
    "pyyaml>=6.0",
    "numpy>=2.0",
    "rich>=14.0",
    "textual>=8.0",
]

[project.optional-dependencies]
local-embedding = [
    "llama-cpp-python>=0.3.0",
]
dev = [
    "pytest>=9.0",
    "ruff>=0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["python"]
